+++ src/h264/sps.c
@@ -361,7 +361,7 @@ void h264_sps_print(const struct h264_sps *sps)
 }
 
 
-const char *h264_sps_chroma_format_name(unsigned chroma_format_idc)
+const char *h264_sps_chroma_format_name(uint8_t chroma_format_idc)
 {
 	switch (chroma_format_idc) {
